refactor: Move tnctest from src into cmd/samoyed-tnctest
Follows the same pattern as the kissutil and walk96 moves: the
CLI-specific logic now lives in cmd/samoyed-tnctest/main.go. No
wrapper file was needed since tnctest only used already-exported
package direwolf symbols (SLEEP_MS, AGWPEHeader, SerialPortOpen,
SerialPortWrite, SerialPortGet1); the column_width/LINE_WIDTH
formatting globals, which used to piggyback on aclients.go's package
vars, are now local to cmd/samoyed-tnctest.
